requirements.txt
python==3.10.8
pytorch==1.13.1
numpy==1.23.5
cudatoolkit==12.0
pip==22.3.1
conda==23.1.0